+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-10-06
+
+### Changed
+
+- Implement powershell build script that can be run from any location
+- Implement powershell build script that searches for latex file in its PSRoot folder
+- Implement powershell script that renames main latex and bibtex files to match repo root directory name
+- Update README with new build instructions
+
 ## [1.2.0] - 2024-09-25
 
 ### Changed